Refactor Dialog and DialogManager for improved key generation
The previous implementation of Dialog and DialogManager was refactored to improve key generation. Instead of using just chatId, it is now combined with userId to generate a unique key for a dialog. This solution further enhance balance between performance and security for reading and storing dialog states. Some unnecessary methods and properties were also cleaned up in the process.